pyproject.toml
[tool.poetry]
name = "adventofcode2021"
version = "0.1.0"
description = ""
authors = ["Madelaine Boyd <madelaine.boyd+github@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.8, <3.11"
pymp-pypi = "^0.4.5"
numpy = "^1.21.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
